Add files via upload
Python application that can fetch, process, and display 
weather data for different cities. The application should allow users to input a city 
and retrieve its current weather data, as well as historical weather data for the last 
seven days. The user should be able to perform simple statistical analysis on the 
historical data and save the results to a file.